+++ examples/lib/main.dart
@@ -14,6 +14,8 @@ import 'package:example/games/games_fps2.dart';
 import 'package:example/geometry/marching_cubes.dart';
 import 'package:example/geometry/webgl_buffergeometry_custom_attributes_particles.dart';
 import 'package:example/geometry/webgl_decals.dart';
+import 'package:example/geometry/webgl_decals2.dart';
+import 'package:example/geometry/webgl_geometries_parametric.dart';
 import 'package:example/geometry/webgl_geometry_convex.dart';
 import 'package:example/geometry/webgl_geometry_dynamic.dart';
 import 'package:example/geometry/webgl_geometry_extrude_shapes.dart';
@@ -31,6 +33,9 @@ import 'package:example/instancing/webgl_instancing_raycasting.dart';
 import 'package:example/instancing/webgl_instancing_scatter.dart';
 import 'package:example/lights/webgl_lightprobe.dart';
 import 'package:example/lights/webgl_lightprobe_cube_camera.dart';
+import 'package:example/line/webgl_interactive_lines.dart';
+import 'package:example/line/webgl_lines_colors.dart';
+import 'package:example/line/webgl_lines_dashed.dart';
 import 'package:example/line/webgl_lines_fat.dart';
 import 'package:example/line/webgl_lines_fat_raycasting.dart';
 import 'package:example/line/webgl_lines_fat_wireframe.dart';
@@ -49,6 +54,7 @@ import 'package:example/loaders/webgl_loader_stl.dart';
 import 'package:example/loaders/webgl_loader_usdz.dart';
 import 'package:example/loaders/webgl_loader_vox.dart';
 import 'package:example/loaders/webgl_loader_xyz.dart';
+import 'package:example/material/webgl_materials_video.dart';
 import 'package:example/src/statistics.dart';
 import 'package:example/texture/webgl_video_texture.dart';
 import 'package:example/material/webgl_materials_car.dart';
@@ -275,6 +281,15 @@ class MyAppState extends State<MyApp> {
               '/webgl_interactive_points':(BuildContext context) {
                 return const WebglInteractivePoints();
               },
+              '/webgl_interactive_lines':(BuildContext context) {
+                return const WebglInteractiveLines();
+              },
+              '/webgl_lines_colors':(BuildContext context) {
+                return const WebglLinesColors();
+              },
+              '/webgl_lines_dashed':(BuildContext context) {
+                return const WebglLinesDashed();
+              },
               '/webgl_lines_fat':(BuildContext context) {
                 return const WebglLinesFat();
               },
@@ -286,6 +301,9 @@ class MyAppState extends State<MyApp> {
               },
               '/webgl_decals':(BuildContext context) {
                 return const WebglDecals();
+              },
+              '/webgl_decals2':(BuildContext context) {
+                return const WebglDecals2();
               },
               '/webgl_sprites':(BuildContext context) {
                 return const WebglSprites();
@@ -334,6 +352,9 @@ class MyAppState extends State<MyApp> {
               },
               '/webgl_geometry_spline_editor':(BuildContext context) {
                 return const WebglGeometrySplineEditor();
+              },
+              '/webgl_geometries_parametric':(BuildContext context) {
+                return const WebglGeometriesParametric();
               },
               '/three_terrain':(BuildContext context) {
                 return const TerrainPage();
@@ -445,6 +466,9 @@ class MyAppState extends State<MyApp> {
               },
               '/flutter_material':(BuildContext context) {
                 return const WebglVideoTexture();
+              },
+              '/webgl_materials_video':(BuildContext context) {
+                return const WebglMaterialsVideo();
               },
               '/webgl_materials_car':(BuildContext context) {
                 return const WebglMaterialsCar();