+++ packages/three_js/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.1.2
+
+* Updated three_js_core to fix material bug
+* Added VideoTexture for web only
+
+## 0.1.1
+
+* Updated three_js_core to fix android bug.
+
 ## 0.1.0+1
 
 * Updated readme.